+++ src/test/java/io/github/opencubicchunks/cubicchunks/cubicgen/TestCustomGeneratorSettingsUserFunction.java
@@ -25,16 +25,30 @@ package io.github.opencubicchunks.cubicchunks.cubicgen;
 
 import static org.junit.Assert.assertEquals;
 
+import com.flowpowered.noise.Noise;
 import io.github.opencubicchunks.cubicchunks.cubicgen.customcubic.CustomGeneratorSettings;
+import io.github.opencubicchunks.cubicchunks.cubicgen.customcubic.builder.IBuilder;
+import io.github.opencubicchunks.cubicchunks.cubicgen.customcubic.builder.NoiseSource;
+import net.minecraft.util.math.MathHelper;
+import net.minecraft.world.gen.NoiseGeneratorPerlin;
 import org.junit.Test;
+import org.spongepowered.noise.LatticeOrientation;
+import org.spongepowered.noise.module.source.Simplex;
 
+import java.awt.image.BufferedImage;
+import java.io.File;
+import java.io.IOException;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Random;
+
+import javax.imageio.ImageIO;
 
 public class TestCustomGeneratorSettingsUserFunction {
 
     @Test public void testNoValues() {
-        Map<Float, Float> funcMap = new HashMap<>();
+        Map<Double, Double> funcMap = new HashMap<>();
         CustomGeneratorSettings.UserFunction func = new CustomGeneratorSettings.UserFunction(funcMap);
 
         for (int y = -100; y < 100; y++) {
@@ -43,19 +57,19 @@ public class TestCustomGeneratorSettingsUserFunction {
     }
 
     @Test public void testSingleValue() {
-        Map<Float, Float> funcMap = new HashMap<>();
-        funcMap.put(1.23456789f, 9.87654321f);
+        Map<Double, Double> funcMap = new HashMap<>();
+        funcMap.put(1.23456789, 9.87654321);
         CustomGeneratorSettings.UserFunction func = new CustomGeneratorSettings.UserFunction(funcMap);
 
         for (int y = -100; y < 100; y++) {
-            assertEquals(9.87654321f, func.getValue(y), 0);
+            assertEquals(9.87654321, func.getValue(y), 0);
         }
     }
 
     @Test public void test2Values() {
-        Map<Float, Float> funcMap = new HashMap<>();
-        funcMap.put(0.0f, 0.0f);
-        funcMap.put(1.0f, 1.0f); // identity function
+        Map<Double, Double> funcMap = new HashMap<>();
+        funcMap.put(0.0, 0.0);
+        funcMap.put(1.0, 1.0); // identity function
         CustomGeneratorSettings.UserFunction func = new CustomGeneratorSettings.UserFunction(funcMap);
 
         for (int y = -100; y < 100; y++) {
@@ -65,10 +79,10 @@ public class TestCustomGeneratorSettingsUserFunction {
     }
 
     @Test public void test3Values() {
-        Map<Float, Float> funcMap = new HashMap<>();
-        funcMap.put(-1.0f, 1.0f);
-        funcMap.put(0.0f, 0.0f);
-        funcMap.put(1.0f, 1.0f); // abs function
+        Map<Double, Double> funcMap = new HashMap<>();
+        funcMap.put(-1.0, 1.0);
+        funcMap.put(0.0, 0.0);
+        funcMap.put(1.0, 1.0); // abs function
         CustomGeneratorSettings.UserFunction func = new CustomGeneratorSettings.UserFunction(funcMap);
 
         for (int y = -10; y < 100; y++) {
